requirements.txt
Django==2.2.6
djangorestframework==3.9.3
django-memcache-status==2.1
django-filters==0.2.1
django-filter==2.2.0
object_mapper